requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langchain-text-splitters>=0.3.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.2.0
pgvector>=0.3.0
apache-airflow>=2.9.0
google-genai>=1.0.0
python-telegram-bot>=21.0
chromadb>=0.5.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.0.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
duckduckgo-search>=4.0.0
readability-lxml>=0.8.1
markdownify>=0.11.0
trafilatura>=2.0.0
feedparser>=6.0.0
nodriver>=0.38
crawl4ai>=0.8.0
pytest>=7.0.0
pytest-asyncio>=0.23.0
aiohttp>=3.9.0
mcp>=1.0.0
tavily-python>=0.5.0